feat: new TSObjectFactory API to allow users to provide custom implementaions of TS objects
BREAKING CHANGE: This change makes constructors of all TS objects protected so that they can only be accessed within the same package and in subclasses. Every TS object now provides
a static 'create' method to create instances of the same. Users are expected to use those methods or TSObjectFactory directly.